1. 重构项目模块结构，修复若干bug 2. 重新调整项目整体的异常拦截，异常返回提示信息更加丰富 3. 更新登录页面，增加锁屏界面 4. 增加在线数据源配置功能，使用多数据源更加方便 5. 升级核心包版本到v2.2.12 6. 升级easyweb框架版本v3.1.7 7. 完善通知页面的富文本框，采用ueditor 8. 更新主题界面样式

+++ guns-base-support/guns-base-auth/src/main/java/cn/stylefeng/guns/base/auth/jwt/payload/JwtPayLoad.java
@@ -65,7 +65,7 @@ public class JwtPayLoad {
             JwtPayLoad jwtPayLoad = new JwtPayLoad();
 
             Object userId = map.get("userId");
-            if (userId instanceof Integer) {
+            if (userId instanceof Long) {
                 jwtPayLoad.setUserId(Long.valueOf(map.get("userId").toString()));
             }
